refactor(metactl): extract Lua runtime setup to dedicated module
Move Lua gRPC client functionality from main.rs to a new lua_support
module in databend-common-meta-control. This improves code organization
by separating Lua runtime concerns from the main binary logic.

Changes:
- Create src/meta/control/src/lua_support.rs with LuaGrpcClient and setup functions
- Update main.rs to use lua_support::setup_lua_environment()
- Add mlua and meta-kvapi dependencies to control library
- Remove duplicate Lua implementation code from main.rs

This refactoring makes the Lua functionality more maintainable and
reusable across the codebase while preserving all existing behavior.
