+++ src/meta/control/src/lib.rs
@@ -20,5 +20,6 @@ pub mod export_from_disk;
 pub mod export_from_grpc;
 pub mod import;
 mod import_v004;
+pub mod lua_support;
 pub(crate) mod reading;
 mod upgrade;
